Move Battle-related fixtures to fixtures dir
Additional changes:
- Write tests to verify Building multipliers successfully apply to BattleToons

Up next:
- Verify Toons retrieve correct reward values when selecting attacks
