+++ tests/battle/conftest.py
@@ -1,2 +1,4 @@
 from ..fixtures.battle_fixtures import (  # noqa to suppress unused import
-    bc_random, bc_lured, bc_trapped, bt_astro, bt_trapa)
+    bc_lured, bc_random, bc_trapped, bt_astro, bt_trapa, building_floor,
+    expected_building_multiplier, battle_building
+)
