+++ tests/fixtures/battle_fixtures.py
@@ -1,9 +1,15 @@
 import pytest
 
-from ...Battle import BattleToon
+from ...AttackGlobals import MULTIPLIER
+from ...Battle import Battle, BattleToon, RewardCalculator
 from ...Cog import Cog, get_random_cog
 from ...Factory import BattleCogFactory
 from ...Toon import Toon
+
+
+@pytest.fixture
+def battle_building(building_floor: int) -> Battle:
+    return Battle(building_floor=building_floor)
 
 
 @pytest.fixture
@@ -68,6 +74,18 @@ def bc_trapped(request):
                                            trapped=True)
 
 
+@pytest.fixture(params=[1, 2, 3, 4, 5], scope='module')
+def building_floor(request) -> int:
+    """Return all possible building floor values, one at a time"""
+    return request.param
+
+
+@pytest.fixture(scope='module')
+def expected_building_multiplier(building_floor: int) -> float:
+    """Given a building_floor value, return the expected building_multiplier value"""
+    return MULTIPLIER.get_building_multiplier_from_floor(floor=building_floor)
+
+
 def get_cog_from_request_param(request_param) -> Cog:
     """Given a pytest request containing a Cog or a (key, rel_lvl) tuple, return a Cog object
 
@@ -83,3 +101,8 @@ def get_cog_from_request_param(request_param) -> Cog:
         return request_param
     else:
         raise TypeError(f"What the heck did you request? {request_param}")
+
+
+def get_reward_calculator(building_floor: int = 1, is_invasion: bool = False) -> RewardCalculator:
+    """Return a RewardCalculator, given a building_floor number and is_invasion boolean"""
+    return RewardCalculator(building_floor=building_floor, is_invasion=is_invasion)
